+++ generators/app/templates/es/jest.config.js
@@ -27,7 +27,6 @@ module.exports = {
   testURL: 'http://localhost',
   transform: {
     '^.+\\.(js|jsx|mjs)$': 'babel-jest',
-    '^.+\\.md?$': 'markdown-loader-jest',
     '^.+\\.s?css$': '<rootDir>/config/jest/style-transform.js',
     '^.+\\.(png|jpg|jpeg|git|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
       '<rootDir>/config/jest/file-transform.js'