+++ generators/app/templates/es/src/routes/app-routes.js
@@ -2,16 +2,11 @@
  * @fileoverview Defines top-level application routes.
  */
 import Home from '@src/components/home';
-import Documentation from '@src/components/documentation';
 
 const routes = [
   {
     path: '/',
     component: Home
-  },
-  {
-    path: '/documentation',
-    component: Documentation
   }
 ];