+++ generators/app/templates/ts/scripts/test.ts
@@ -4,7 +4,6 @@
  */
 import * as jest from 'jest';
 
-process.env.BABEL_ENV = 'test';
 process.env.NODE_ENV = 'test';
 
 process.on('unhandledRejection', (err) => {