+++ generators/app/templates/ts/src/components/home/Home.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import Layout from '@src/components/shared/layout';
 import SectionHeader from '@shared/section-header';
 
@@ -7,7 +6,7 @@ const Home: React.FC = () => (
   <Layout>
     <SectionHeader title="Welcome to your new React App!" description="Made with ❤️." />
     <p>
-      <Link to="/documentation">Click here</Link> to view the build instructions.
+      Reference <code>README.md</code> for build instructions.
     </p>
   </Layout>
 );