+++ generators/app/templates/ts/src/routes/app-routes.ts
@@ -3,16 +3,11 @@
  * @packageDocumentation
  */
 import Home from '@src/components/home';
-import Documentation from '@src/components/documentation';
 
 const routes = [
   {
     path: '/',
     component: Home
-  },
-  {
-    path: '/documentation',
-    component: Documentation
   }
 ];